fix label

+++ LabellingTool.py
@@ -3,23 +3,26 @@ import os
 import csv
 import argparse
 
+import cv2
 from PySide6 import QtCore
 from PySide6.QtCore import QDir, Qt, QRectF, QPoint, QPointF
 from PySide6.QtGui import QImage, QKeyEvent, QPainter, QPixmap, QColor, QPen, QFont, QBrush, QTransform
 from PySide6.QtWidgets import QApplication, QLabel, QMainWindow, QGraphicsScene, QGraphicsView, QGraphicsEllipseItem, QWidget
 from PySide6.QtWidgets import QHBoxLayout, QVBoxLayout, QPushButton, QGroupBox
 from PySide6.QtWidgets import QFileDialog, QGraphicsTextItem
-from PySide6.QtWidgets import QApplication, QGraphicsView, QGraphicsScene, QGraphicsPixmapItem
+from PySide6.QtWidgets import QApplication, QGraphicsView, QGraphicsScene, QGraphicsPixmapItem, QCheckBox
 
 parser = argparse.ArgumentParser(description='Argument Parser for GridTrackNet')
 
 parser.add_argument('--match_dir', required=True, type=str, help="Match directory. Must be named with prefix 'match' following by an index.")
+parser.add_argument('--video_path', type=str, default=None, help="Optional path to a video file. If provided, frames are read on demand via OpenCV instead of match_dir/frames.")
 
 args = parser.parse_args()
 
 MATCH_DIR = args.match_dir
 FRAMES_DIR = os.path.join(MATCH_DIR, "frames")
 CSV_DIR = os.path.join(MATCH_DIR, "Labels.csv")
+VIDEO_PATH = args.video_path
 
 if(not os.path.exists(MATCH_DIR)):
     print("\nERROR: The following directory does not exist: " + str(MATCH_DIR))
@@ -36,7 +39,11 @@ if(not (validDir(MATCH_DIR))):
     print("\nERROR: Specified export folder does not start with 'match' followed by an index.")
     exit(1)
 
-if(not os.path.exists(FRAMES_DIR)):
+if VIDEO_PATH is not None and not os.path.exists(VIDEO_PATH):
+    print("\nERROR: The following video file does not exist: " + str(VIDEO_PATH))
+    exit(1)
+
+if(VIDEO_PATH is None and not os.path.exists(FRAMES_DIR)):
     print("\nERROR: The following directory does not exist: " + str(FRAMES_DIR))
     exit(1)
 
@@ -55,14 +62,19 @@ class ImageViewer(QMainWindow):
         self.view.mousePressEvent = self.getPixelCoordinates
         
         self.images = []
+        self.frameLabels = []
+        self.video_path = VIDEO_PATH
+        self.video_capture = None
+        self.video_frame_count = 0
         self.frameIndex = 0
         self.pixelCoordinates = {}
         self.states = {}
         self.annotated = {}
+        self.showCrosshair = True
         self.currentCenterPoint = self.view.mapToScene(self.view.viewport().rect().center()) 
         self.loadImages()
 
-        self.pixmap = QPixmap(self.images[self.frameIndex])
+        self.pixmap = self.loadFramePixmap(self.frameIndex)
         self.scene = QGraphicsScene()
         self.scene.addPixmap(self.pixmap)
         self.view.setScene(self.scene)
@@ -85,6 +97,11 @@ class ImageViewer(QMainWindow):
         self.saveResultsButton.setFixedSize(200, 75)
         self.saveResultsButton.clicked.connect(self.saveResults)
 
+        self.crosshairCheckbox = QCheckBox("Show Crosshair", self)
+        self.crosshairCheckbox.setChecked(True)
+        self.crosshairCheckbox.setFont(QFont("Arial", 20))
+        self.crosshairCheckbox.stateChanged.connect(self.toggleCrosshairVisibility)
+
         self.visbilityText = QLabel()
         self.visbilityText.setStyleSheet("color: red;")
         self.visbilityText.setFont(QFont("Arial", 30))
@@ -106,6 +123,8 @@ class ImageViewer(QMainWindow):
         self.buttonLayout.addWidget(self.removePixelButton)
         self.buttonLayout.addWidget(self.removeFrameButton)
         self.buttonLayout.addWidget(self.saveResultsButton)
+        self.buttonLayout.addWidget(self.crosshairCheckbox)
+        self.buttonLayout.addStretch()
 
         self.bottomLayout = QHBoxLayout()
         self.bottomLayout.addWidget(self.view)
@@ -139,27 +158,80 @@ class ImageViewer(QMainWindow):
         
 
     def loadImages(self):
-        global FRAMES_DIR 
-        directory = QDir(FRAMES_DIR)
-        directory.setNameFilters(["*.png"])
-        directory.setSorting(QDir.Name)
-        self.images = [directory.filePath(file) for file in directory.entryList()]
-        self.images = sorted(self.images, key=lambda x: int(x.split('/')[-1].split(".")[0]))
+        global FRAMES_DIR
+
+        if self.video_path is not None:
+            self.video_capture = cv2.VideoCapture(self.video_path)
+            if not self.video_capture.isOpened():
+                print("\nERROR: Failed to open video: " + str(self.video_path))
+                exit(1)
+
+            self.video_frame_count = int(self.video_capture.get(cv2.CAP_PROP_FRAME_COUNT))
+            self.images = list(range(self.video_frame_count))
+            self.frameLabels = [f"frame_{index}" for index in self.images]
+        else:
+            directory = QDir(FRAMES_DIR)
+            directory.setNameFilters(["*.png"])
+            directory.setSorting(QDir.Name)
+            self.images = [directory.filePath(file) for file in directory.entryList()]
+            self.images = sorted(self.images, key=lambda x: int(x.split('/')[-1].split(".")[0]))
+            self.frameLabels = [path.split('/')[-1] for path in self.images]
+
+        if len(self.images) == 0:
+            print("\nERROR: No frames were found for annotation.")
+            exit(1)
 
         for i in range (len(self.images)):
             self.annotated[i] = False
 
+    def loadFramePixmap(self, frame_index):
+        if self.video_path is None:
+            return QPixmap(self.images[frame_index])
+
+        source_frame_index = self.images[frame_index]
+        self.video_capture.set(cv2.CAP_PROP_POS_FRAMES, source_frame_index)
+        success, frame = self.video_capture.read()
+        if not success:
+            raise RuntimeError(f"Failed to read frame {source_frame_index} from {self.video_path}")
+
+        frame_rgb = cv2.cvtColor(frame, cv2.COLOR_BGR2RGB)
+        height, width, channels = frame_rgb.shape
+        bytes_per_line = channels * width
+        image = QImage(frame_rgb.data, width, height, bytes_per_line, QImage.Format_RGB888).copy()
+        return QPixmap.fromImage(image)
+
+    def getFrameDisplayName(self, frame_index):
+        return self.frameLabels[frame_index]
+
+    def drawCrosshair(self, x, y):
+        if not self.showCrosshair:
+            return
+
+        crosshair_pen = QPen(Qt.red)
+        crosshair_pen.setWidth(2)
+        arm_length = 18
+        gap_size = 6
+
+        self.view.scene().addLine(x - gap_size - arm_length, y, x - gap_size, y, crosshair_pen)
+        self.view.scene().addLine(x + gap_size, y, x + gap_size + arm_length, y, crosshair_pen)
+        self.view.scene().addLine(x, y - gap_size - arm_length, x, y - gap_size, crosshair_pen)
+        self.view.scene().addLine(x, y + gap_size, x, y + gap_size + arm_length, crosshair_pen)
+
+    def toggleCrosshairVisibility(self, state):
+        self.showCrosshair = state == int(Qt.CheckState.Checked)
+        self.showImage()
+
     def showImage(self):
         self._apply_render_hints()
 
-        self.pixmap = QPixmap(self.images[self.frameIndex])
+        self.pixmap = self.loadFramePixmap(self.frameIndex)
         self.view.setScene(QGraphicsScene())
         self.view.scene().addPixmap(self.pixmap)
         self.view.fitInView(self.view.sceneRect(), Qt.KeepAspectRatio)
         self.view.scale(self.zoomLevel, self.zoomLevel)
         self.view.centerOn(self.currentCenterPoint)
 
-        self.imageText.setText(self.images[self.frameIndex].split('/')[-1])
+        self.imageText.setText(self.getFrameDisplayName(self.frameIndex))
 
         if not self.frameIndex in self.states:
             self.states[self.frameIndex] = "VISIBLE"
@@ -176,8 +248,7 @@ class ImageViewer(QMainWindow):
 
         if(self.frameIndex in self.pixelCoordinates):
             x, y = self.pixelCoordinates[self.frameIndex]
-            dotSize = 1
-            self.view.scene().addEllipse(x-dotSize/2, y-dotSize/2, dotSize, dotSize, QPen(Qt.red))
+            self.drawCrosshair(x, y)
    
     def toggleState(self):
         if(self.frameIndex in self.states):
@@ -201,7 +272,7 @@ class ImageViewer(QMainWindow):
             self.showImage()
 
     def removeFrame(self):
-        print("Removed image: " + str(self.images[self.frameIndex]))
+        print("Removed image: " + str(self.getFrameDisplayName(self.frameIndex)))
 
         if(self.frameIndex in self.pixelCoordinates):
             self.pixelCoordinates.pop(self.frameIndex)
@@ -224,9 +295,16 @@ class ImageViewer(QMainWindow):
                     value = self.annotated.pop(key)
                     self.annotated[key-1] = value
         
-        os.remove(self.images[self.frameIndex])
+        if self.video_path is None:
+            os.remove(self.images[self.frameIndex])
         
         self.images.pop(self.frameIndex)
+        self.frameLabels.pop(self.frameIndex)
+
+        if len(self.images) == 0:
+            print("All frames were removed.")
+            self.close()
+            return
 
         if(self.frameIndex >= len(self.images)):
             self.frameIndex = len(self.images) - 1
@@ -241,17 +319,16 @@ class ImageViewer(QMainWindow):
         for i in range(len(self.images)):
             if(not self.annotated[i]):
                 allAnnotated = False
-                print(str(self.images[i].split('/')[-1]) + " is not annotated.")
+                print(str(self.getFrameDisplayName(i)) + " is not annotated.")
 
         if(allAnnotated):
             with open(CSV_DIR, "w") as f:
                 writer = csv.writer(f)
                 writer.writerow(["Frame", "Visibility", "X", "Y"])
-                indices = list(range(0,len(self.pixelCoordinates)))
     
                 data = []   
                 for i in range(len(self.images)):
-                    if(self.states[i] == "VISIBLE"):
+                    if(self.states.get(i, "VISIBLE") == "VISIBLE"):
                         visibility = 1
                         
                     else:
@@ -264,25 +341,36 @@ class ImageViewer(QMainWindow):
                         x_coord = 0
                         y_coord = 0
                         
-                    data.append([i, visibility, x_coord, y_coord])
+                    source_index = self.images[i] if self.video_path is not None else i
+                    data.append([source_index, visibility, x_coord, y_coord])
 
                 writer.writerows(data)
                 print("Saved Results!")
             
-            for i in range(len(self.images)):
-                os.rename(str(self.images[i]), os.path.join(FRAMES_DIR, str(i) + ".png"))
+            if self.video_path is None:
+                for i in range(len(self.images)):
+                    os.rename(str(self.images[i]), os.path.join(FRAMES_DIR, str(i) + ".png"))
 
-            sys.exit(app.exec())
+            QApplication.instance().quit()
+
+    def stepFrame(self, offset):
+        if len(self.images) == 0:
+            return
+
+        new_index = max(0, min(len(self.images) - 1, self.frameIndex + offset))
+        if new_index != self.frameIndex:
+            self.frameIndex = new_index
+            self.showImage()
       
     def keyPressEvent(self, event: QKeyEvent):
         if event.key() == Qt.Key_A:
-            if self.frameIndex > 0:
-                self.frameIndex -= 1
-                self.showImage()
+            self.stepFrame(-1)
         elif event.key() == Qt.Key_D:
-            if self.frameIndex < len(self.images) - 1:
-                self.frameIndex += 1
-                self.showImage()
+            self.stepFrame(1)
+        elif event.key() == Qt.Key_W:
+            self.stepFrame(15)
+        elif event.key() == Qt.Key_S:
+            self.stepFrame(-15)
 
     def getPixelCoordinates(self, event):
         pos = event.pos()
@@ -291,8 +379,7 @@ class ImageViewer(QMainWindow):
         self.pixelCoordinates[self.frameIndex] = (scene_pos.x(), scene_pos.y())
         self.annotated[self.frameIndex] = True
 
-        dotSize = 1
-        self.view.scene().addEllipse(scene_pos.x()-dotSize/2, scene_pos.y()-dotSize/2, dotSize, dotSize, QPen(Qt.red))
+        self.drawCrosshair(scene_pos.x(), scene_pos.y())
         
         self.currentCenterPoint = QPointF(scene_pos.x(), scene_pos.y())
         if self.frameIndex < len(self.images) - 1:
